Extract index operations into index module
